requirements.txt
python-telegram-bot[job-queue]==20.8
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
pytz==2024.1
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
oauth2client==4.1.3
gspread==6.0.2
httpx>=0.26.0  # For better HTTP/2 support
